+++ src/Providers/SeoServiceProvider.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Honeystone\Seo\Providers;
 
+use Honeystone\Seo\Commands\GenerateFavicons;
 use Honeystone\Seo\Contracts\BuildsMetadata;
 use Honeystone\Seo\Contracts\RegistersGenerators;
 use Honeystone\Seo\MetadataDirector;
@@ -23,6 +24,7 @@ final class SeoServiceProvider extends PackageServiceProvider
         $package
             ->name('honeystone-seo')
             ->setBasePath(dirname(__DIR__))
+            ->hasCommand(GenerateFavicons::class)
             ->hasConfigFile()
             ->hasViews();
     }