+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.3.0 - 2024-08-21
+
+- New favicon generator using the [realfavicongenerator.net API](https://realfavicongenerator.net/api/non_interactive_api).
+
+## 1.2.0 - 2024-08-02
+
+- Added basic Inertia and Vue 3 integration.
+
 ## 1.1.0 - 2024-07-27
 
 - JSON-LD configured with this package will be automatically appended to the Graph or MultiTypedEntity when using the